Allow navigation back to 'home' page. Add carousel folders view.
Change the home page so we can navigate back to it.  The home page
used to be the place where we'd jump to ask for password or to the
list of folders.  Now we have an intermediate page that can direct
us to the proper start up page.

Added a simple implementation of the carousel just so our fingers
can get used to it.

Added the iCarousel project to the NachoClientX project.