Modify a forwarded message that has attachments.
Handle the case where a message with attachments is forwarded, and the user edits the body of the forwarded message. If all of the attachments have already been downloaded, then automatically attach them to the outgoing message. If any attachments need to be downloaded, ask the user whether or not to send the message without the attachments.

(This is a temporary fix. In the future, this will be changed to automatically download the necessary attachments before sending the message.)