Convert smart F/R to send; un-downloaded attachments during forward
Provide a method that converts a message from a smart forward or smart reply to a regular send. This conversion may involve adding the body of the original message into the body of the outgoing message, and adding any attachments from the original message.

Correctly handle the case where a message with attachments is forwarded, but some of those attachments have not yet been downloaded.  The attachments need to be downloaded before they can be attached to the outgoing message.