Improve code to update an existing McBody
Provide a way to update the existing contents of a McBody when you don't have a string, but instead have code that writes to a FileStream.

Improve the code for updating a McBody so that files are better managed and the backing file spends less time in an inconsistent state.