Get attachments on calendar events mostly working.
When the event is created on the device, attachments cannot be synched with the server. So we don't even try. The attachments will still be visible on the device, and they will be visible to anyone invited to the meeting. They just aren't visible to other clients of the same account.

Most of the testing was of creating appointments and non-repeating meetings. That works well, except for the above mentioned limitation. Editing existing events has problems. The attachment viewer when viewing events doesn't work. These problems will be fixed later.