Three bug fixes for calendar sync. Issues #518 and #520.
Stop sending the ResponseType element of events in command requests. That field should only be in responses.  (Section 2.2.2.38 in MS-ASCAL.)

Stop sending the AttendeeStatus element of Attendee elements in command requests. That field should only be in responses.  (Microsoft's documentation is unclear and incorrect for this element.)

For the UserResponse element of the MeetingResponse command, send the integer value of the enum (e.g. "1") rather than the symbolic name of the enum constant (e.g. "Accepted_1"). This was causing all MeetingResponse commands to fail.