Fix bugs in display of meeting response status.
The button bar at the bottom of the event view was looking for the user's status in the attendee section of the meeting.  That's the wrong place to look.  The user's meeting status is in the ResponseType field of the event.

Clean up the code for displaying the meeting attendees in the event view.  It was poorly organized and had a big block of duplicated code.

The ResponseType field of McAbstrCalendarRoot was not being saved in the database, because it was a field rather than a property.