Send meeting response e-mail when responding to a meeting request
When the user accepts/declines a meeting from the calendar, and the organizer requested response, then a meeting response e-mail message will be sent to the organizer. This allows the user's status to be recorded in the organizer's copy of the meeting.

The user is not given the option to not send the message, or to add some custom text to the message. Those features will be covered later (if at all).

This e-mail message is not sent when responding to the meeting request from the inbox.  Meeting requests in the inbox have more problems than just the response e-mail message, so that will be covered in a separate update.