Issue #517 Display calendar event descriptions that aren't plain text.
Until now, calendar event descriptions would only be displayed in the event
viewer if a plain text version were available.  But a plain text version of the
description is usually not available, so most events showed an empty
description.

To fix this, EventViewController was changed to use a BodyView instead of an
ordinary UILabel for displaying the description.  Body View had to be enhanced
to handle TNEF sections and RTF (since the Exchange server will agressively
reformat event descriptions into those formats).