Fix display of HTML calendar event descriptions.
The detail view for events with HTML descriptions was causing a crash.  (HTML
event descriptions are rare, which is why this bug was not noticed yesterday.
The only way I can get one is to cancel a meeting.) Fix the code to not crash
and to display the description with the correct indentation.