Fix event view scroll problem.
Event views where the description is RTF (which is most of them) would
often have problems scrolling. If the user touched within the
description, iOS would try to scroll just the description rather than
the entire view, but that would have no effect because the description
is entirely visible (within its outer scroll view). To scroll the
entire event view, the user would have to touch outside the
description (which could be difficult if the description filled the
entire screen).

The problem is a bug in iOS. It handles nested UIScrollViews just fine
as long as they are only nested two deep. When a third UIScrollView is
added to the hierarchy, iOS behaves badly.

The fix is to disable scrolling in the innermost UITextView, which is
derived from UIScrollView.  There are still three nested UIScrollView
objects in the hierarchy, but only two of them are potentially
scrollable, thus avoiding the iOS bug.