Handle meeting response and meeting cancellation messages.
Do the right thing with meeting response and meeting cancellation
messages.  The code for displaying the details of the message is
shared with meeting requests and is mostly unchanged.  What is
different is the action bar.  Instead of three buttons for Accept,
Tentative, and Decline, meeting responses display the attendee's
response (with no actions available), and meeting cancellations have a
"Remove from calendar" button.

This update focuses on correct functionality and on displaying the
correct information.  It does not focus on getting the UI exactly
right.  That will be done separately.