Don't crash when there are events with duplicate UIDs.
We are seeing cases where databases have duplicate entries for events,
where the two entries have the same UID but different ServerIds.  This
was causing an assertion failure in McCalendar.QueryByUID() when
viewing a meeting request message for one of these events.

Change McCalendar.QueryByUID() to not crash when it finds duplicate
events. Instead it will log an error message and return the most
recently modified of the matching event records.

Change AsSyncCommand.ServerSaysAddOrChangeCalendarItem() to log an
error message when it detects something fishy with the UID or ServerId
for an event.  This change will not stop duplicate records from being
created.  (Since we don't know what is triggering the duplicate
records, we don't know the proper way to stop them.)  But the error
messages will hopefully help us find the cause of this problem.

Also fix AsSyncCommand.ServerSaysAddOrChangeCalendarItem() so it won't
create McCalendar items with a CreatedAt field with the value 0.