Don't cut off the bottom of messages with attachments.
When a message had attachments and was longer than one screen, the
message detail view would cut off the very bottom of the message. The
problem was that one of the height calculations was using the location
of the separator above the attachments section instead of the
separator below the attachments section.