Improve time zone support and handling of all-day events.
Better handling of all-day events, both creating them and displaying
them.  All-day events that are created in Nacho Mail now run from
midnight to midnight (local time), as they are supposed to.  All-day
events no longer spill over into an extra day, but are shown only on
the day (or days) that they are supposed to, even if they were created
in a different time zone.  (The display of all-day events was only
fixed in the calendar view.  They are still not quite right in the
event detail view or in meeting request messages for all-day
meetings.)

Improve the time zone support.  The time zone information baked into
the event that comes from Exchange is now parsed.  (Currently it is
only used for all-day events. Using it for recurring events will come
later, but the necessary infrastructure is there.)  The time zone
information that Nacho Mail inserts into calendar events that it
creates is more accurate.