When a local notification starts the app, iOS delivers it to FinshedLaunching and ReceivedLocalNotification and the app is considered Active.
When a local notification is delivered to a running app, it
is delivered to ReceivedLocalNotification and the app is
consider Active too.

When the app is stared because of a local notification we
want to navigate to the event/message but when the app is
already running, we just want to display an alert for the
event.

To differentiate these cases, we a flag in FinishedLaunched
to tell us that the app was just launched.

Nacho Now handles the event and message navigation.  It looks
at the mutables that hold the event/message information.

Note that Nacho Now does not get WillBecomeVisible when the
tab bar becomes visible, so we let the tab bar tell Nacho
Now when to handle the notification.