Improve the display of message bodies and event bodies.
Overhaul BodyView and its related classes. Simplify the design. The
trick of displaying at most one screen worth of a potentially large
view is preserved. The way that downloads of message bodies are
tracked has changed, so that all the tracking happens in BodyView.

Zooming of message bodies is disabled because I couldn't get it
working right. It will be dealt with later.

Fix #421
Fix #234