Fix recurring events.
The code that calculates the instances of recurring events was
rewritten.  The rewrite fixed at least three bugs.  (1) All date/time
calculations are done in the event's original time zone, so times are
now correct when crossing a daylight saving time boundry.  (2) Certain
recurring events could cause the app to go into an infinite loop on
the UI thread. That won't happen any more.  (3) The instances of an
all-day recurring event now have the correct start and end times.

Fix the display of all-day events in the event detail view.  It now
says "all day event" instead of listing the day of the event and the
day after.

Avoid crashing in the event detail view when opening an event with an
empty description that was stored in TNEF format.

Fix #554