Make web views appear to load faster.
Set the initial height of a BodyWebView to be one entire screen rather
than just one point.  This lets the use see the web content as it is
loading.  In the past, the web view remained just one point high (i.e.
completely invisible) until the web page had finished loading.  This
could cause very long delays for large messages before the user saw
anything interesting.  This change doesn't make the web view load any
faster, but it looks a lot faster to the user.

(The problem of loading slowly is not apparent on the simulator.  This
change was tested on an iPhone 4s device, where everything is slow.
The difference was dramatic.)