Avoid crash when message is deleted while downloading its body.
A NullReferenceException could happen if a message was deleted while
it was visible in a message view and its body was in the process of
being downloaded.  (Or at least that's what I think would happen.  I
haven't reproduced the situation.)  Fix this by having BodyView
gracefully deal with the case where the item it is displaying cannot
be found in the database.

Fix #742