Enable zooming of BodyView.
Make it possible to zoom a message body or event body.  Most of the
work happens in the view controller, because it owns the scroll view
where the zooming happens.  Class BodyView needed to be changed to be
aware of zooming because it has to adjust some of its dimensions by
the zoom scale factor.

Message view supports zooming by pinch gesture or by double tap.  When
using a pinch gesture to zoom out, sometimes only part of the body is
shown while the gesture is in progress, but the view fixes itself when
the gesture is complete.

Event view only supports pinch gestures, not double tapping.  The
layout is often messed up while the pinch gesture is in progress, but
it fixes itself when the gesture is complete.