An interrupted download should not be a permanent error.
If the download of a body or an attachment was interrupted by the app
exiting, the command would be marked as failed and the McAbstrFileDesc
for the body or attachment would be set to the state of Error.
Failing the command is the correct thing to do, but setting the file
to Error was wrong, because the download command didn't really fail.

To fix this: Add a new reason to NcResult.WhyEnum:
InterruptedByAppExit.  When a command fails for this reason, set the
FilePresence of the McAbstrFileDesc to None rather than Error.