Don't shut down back end after transition from PerformFetch to foreground
Fix yet one more case of the spinner-of-death. If the app was brought
into the foreground while a PerformFetch operation was in progress,
then a timer was incorrectly left running. When the timer was fired,
it would shut down the back end. This would result in ActiveSync
commands being put on hold. The most obvious symptom of this is that
message body downloads would never finish, leaving the user to stare
at the spinner indefinitely.

The fix is to move the timer from class BackEnd to class AppDelegate,
and to change what the timer does when it fires.  There are also other
improvements in keeping track of the PerformFetch activity.

Fix #858