Reduce number of "Download failed" messages that the user sees
Change BodyView so that it always tries to download a message body
that isn't already downloaded, even if the body's FilePresence is
Error. The "Download failed" error message will be shown to the user
only if BodyView's attempt to download the body fails. Previous
failures to download the body are ignored. BodyView will only try the
download once without the user tapping the "Try again" button, so this
change will not cause an infinite loop of failed download attempts.