Don't order the main McPending Query
Some telemetry data was showing that McPending.Query() was taking
about 150 ms on average, which is much too long.  The SQL was ordering
the results by priority.  But all of the callers of McPending.Query()
were ignoring that order and reordering the results themselves.
Remove the ordering from the SQL, so we don't waste time while holding
a shared lock on the database.