Handle forward/reply of message with empty body
Forwarding or replying to a message with an empty body was causing a
crash.  I consider this a bug in the specification of C#'s
String.Remove.  It should allow an index that refers to the very end
of the string.  But it doesn't, so we have to work around that.

Fix #908