Tweak use of time variance info to avoid McEmailMessage.Update calls
The Brain's time variance information that was stored in
McEmailMessage items was almost never used.  Its only use was in a
query to identify messages that still had a state transition left in
the future.  Keep this use, but stop storing other time variance
information the database.  The purpose of this change is to
(hopefully) reduce the number of calls to McEmailMessage.Update().