No UI-initiated body downloads when app is in the background
Bad things happen when the UI initiates message body downloads while
the app is in the background.  This can result in the back end
shutting down before the download completes, which causes the download
to fail, which causes a "Download failed" message to be shown to the
user in the Now view when the app comes into the foreground.

Change BodyView to not initiate downloads if the app is in the
background.  Instead, BodyView will wait for a notification that the
app is in the foreground and will start the download at that time.

Fix #826