Clean up failed McPendings
If an operation fails, the failed McPending item used to be left
sitting in the database forever.  This would slow down certain
operations.  This update implements a strategy for preventing that
build up.

The status handlers that listen for the completion of downloads of
message bodies and attachments now delete the McPending object if the
download failed.

During the app start up, the McPending table is searched for failed
downloads whose last modification time is more than two days old.
Those McPendings are deleted from the database.