Don't load UIWebView when app is in the background
Stop UIWebViews from loading pages while the app is in the background.
If the app is in the background when a BodyWebView is created, the
loading of the page is not started.  If the app goes into the
background while a page is being loaded, the page loading is
cancelled.  When the app comes into the foreground, any page that was
not started or was interrupted will be restarted.