Rewrite class AsDnsOperation to improve auto-d for GFE
Rewrite class AsDnsOperation.  The class now calls res_query()
directly rather than using DnDns to issue the DNS request.  DnDns is
still used to parse the result of res_query().  The error code from
res_query() is checked and the call is retried if the error code is
TRY_AGAIN.

Create a minimal C wrapper around res_query().  The wrapper gets the
value from h_errno, which is not easily accessible from C#, and
returns it via a reference parameter.  This wrapper is necessary so
that AsDnsOperation can know when to retry the res_query() call.

Fix #807