Make MX query results available to the advanced login page
Store the results of the MX query, which happens during auto-d, in the
ProtoControl object.  This way, they are available to UI, specifically
to the advanced login page that comes up when initial account
authentication fails.  (The advanced login page hasn't been changed
yet to use the information.  That will happen later.)

Remove an AutoDInfoEnum parameter from a bunch of methods.  Explicitly
passing that parameter through multiple layers of methods was
unnecessary.  Storing the value in the ProtoControl object is enough.