Fix the GFE happy path
Fix the Google Apps for Work (GFE) happy path case, where the user
enters the correct credentials and the server has the expected MX DNS
record.  This was broken earlier because the wrong server URL was
being constructed after a successful MX lookup.  That has been fixed,
and the construction of server base URLs has been centralized in class
McServer.