Handle on-boarding when user selects Google and MX lookup fails
Correctly handle the onboarding case where the user selected Google
Apps for Work (GFE) and auto-discovery fails (including a failed MX
lookup, which is normally what works for GFE).  In this case, we want
to use the information that the user provided and go ahead with
m.google.com as the server.  This required changing the auto-d state
machine behavior when auto-d fails.

As a side effect of the change above, we also want to change the
behavior when authentication fails.  When the user selects GFE and
then enters an incorrect domain in the e-mail address, the app will
try m.google.com, which means it will get an authentication failure
rather than a server failure.  In this case of GFE plus auto-d
failure, we want to go to the advanced login page rather than the
credentials page.