Work around a daylight saving time transition bug in Mono
The Mono runtime has a bug in the calculation of the transition day
between daylight saving time and standard time.  Sometimes the
transition happens a week early.  Write our own methods for converting
to and from UTC that check whether or not the bug is present and work
around the bug if it is.

Fix #1156