Better management of local notifications
Manage the scheduling of local notifications.  Because iOS has a limit
on the number of scheduled notifications, we can't just schedule the
notification for an event when the event is first created.

Class LocalNotificationManager manages the scheduling of local
notifications.  It queries the McEvents in the database to find the
ones with upcoming reminders, and schedules the notifications for a
reasonable number of the next reminders.

Fix #1194