Get HTML rather than plain text for calendar events from Exchange 2007
Exchange 2007 servers do not support MIME for the bodies of calendar
events.  (The server fails the sync request if you ask for MIME.)  But
the servers do support formatted bodies.  (Unlike Google servers.)
When synching with an Exchange 2007 server, request that the bodies be
in HTML rather than plain text.  This will result in the app doing a
better job of displaying the event descriptions.