Fix problem of scrolling event descriptions with HTML
When an event with an HTML description was scrolled, the description
would disappear before reaching the top of the screen.  This change
fixes that problem.  The code in
EventViewController.ScrollViewScrolled that was calculating the offset
of the BodyView was using the offset of the BodyView within
eventCardView, but was forgetting about the offset of eventCardView
within scrollView.

This fixes one of the two problems in 1284. The other problem is still
there, so the issue can't be closed yet.