Fix problem of initial size of event descriptions with HTML
When an event has a small HTML description, the initial layout of the
event detail view would get the size of the description wrong, making
it one entire screen high.  This extra white space would not go away
until the user scrolled the view.

The problem was triggered by a misunderstanding of one of the details
of how scroll views work.  I had assumed that
UIWebView.ScrollView.ContentSize.Height was always the natural height
of the rendered HTML, regardless of the UIWebView's frame size.
That's not true. The ContentSize.Height might be larger than the
frame's height, but it will never be smaller than the frame's height.
This meant that the BodyWebView was not being shrunk to its correct
height after the initial layout has completed.  The fix is to
temporarily set the UIWebView's height to 1, so we can find out the
natural height of the HTML content.

This bug was not noticed in message detail view because it is rare to
have content underneath the HTML section.  It was very obvious in the
event detail view because the description is always followed by
something else.

Fix #1284