Fix broken NcEmailTest
A recent change to the processing of message previews was causing the
NcEmailTest suite to fail.  The mock McBody didn't have an associated
file, but the code that processes a synched e-mail message was
expecting the McBody to have a readable file.

To fix the test, change the Body element of the test XML to have a
Preview element instead of a Data element.  This avoids the need for a
McBody entirely, and more closely matches the data that normally comes
from the server in a real sync.

Fix #1323