Fix the response bar in meeting request/response/cancelation messages
Clean up the code in BodyCalendarView that displays the action bar in
the Hot view and the message detail view.  Through several months of
changes, the code had become rotten and was doing the wrong thing in
several situations.  The code has been fixed to always display the
correct content in the action bar, and has been reorganized to be more
clear.

This is only a UI change.  There is no change to functionality.  There
is one glaring behavior bug that was not fixed: the
Attend/Maybe/Decline buttons have no effect when there is no event in
the calendar corresponding to the meeting request.  But that fix is
beyond the scope of this change and will be handled separately.

Fix #1221