Fix message detail response bar when there is no calendar event
When there was a meeting request message in the inbox but no
corresponding calendar event, the Attend/Maybe/Decline buttons in the
message detail view were duds and had no effect.  This update fixes
that problem, and slightly changes the behavior even when there is a
calendar event.

The buttons for a meeting request in the message detail view now
always use the RespondEmail command instead of the RespondCal command.
A side effect of this change is that meeting request message is
deleted by the server once the user responds.

Meeting response messages are now assembled from the iCalendar
information in the meeting request rather than from the McCalendar
object.

(This update only affects the Attend/Maybe/Decline buttons in the
message detail view.  The behavior of the buttons in the event detail
view was not changed.)

Fix #1346