Fix crash during auto-d of kp.com e-mail address.
The code was assuming that an SRV DNS lookup would only return SRV
records.  That was an incorrect assumption.  An SRV DNS lookup of
_autodiscover._tcp.kp.com returns a single CNAME record.  This
resulted in an InvalidCastException and a crash.

Fix the code to not make this assumption.  DNS response records that
are not of the expected type are now ignored.  The change was made for
both SRV and MX lookups.