Log an error if the server creates a message about sync failure
A couple weeks ago we had a bug there calendars were not getting
synched with Exchange 2007 servers, and the server was sending an
e-mail message to the user letting them know about the failure.  That
bug has been fixed, but we want to log an error message if the server
ever does that again, so we can spot the problem right away.

Fix #1338