Don't crash when accepting a meeting or editing a meeting
An iOS limitation makes it difficult to use generics with value type
arguments.  The DDay.iCal code uses generics with value type
arguments.  We need to be careful to avoid calling that DDay.iCal
code.  If we aren't careful, the app will crash when run on a real
device (but not on the simulator).

Fix #1382
Fix #1388