Don't crash when editing GFE meetings
When the GFE server syncs a meeting to the client, it leaves out the
AttendeeType field for required attendees.  Our code was asserting
that the attendee field was always being exlicitly set.  Change the
code that parses attendees to assume that an attendee is required if
there is no AttendeeType field.

Fix #1386
Fix #1215
Fix #1214