Choose the correct calendar when creating a new event
Hotmail creates three calendar folders for each user.  Two of them,
Birthdays and US Holidays, are read-only.  Any attempt to add an event
to those calendars will result in an sync failure.  When creating a
new calendar event, it is important that Nacho Mail choose the default
calendar folder for the event.  The user can change the calendar, but
the initial setting should be the default folder.  Change
EditEventViewController to find the default calendar folder among the
list of possible calendar folders when a new event is created.

Fix #1370