Tweak the error messages about calendar event UIDs
If an event arrives from the server without a UID, log an error
message only if the event is a meeting.  Previously, an error message
was also logged when the event was an appointment.

When looking up a calendar event by UID, the lookup is now account
specific.  It was realized that it is normal for the same meeting to
show up in two different accounts on the same device.  The error
message about duplicate UIDs was changed to only be logged if the
events are in the same account.  Previously, the error was logged for
messages in different accounts.

Fix #1278