New events should have a status of "Busy" rather than "Free"
Nacho Mail was not setting the BusyStatus field on newly created
events.  Some servers would treat the meeting as if the user were
"Free", while others would treat it as "Busy".  There needs to be
consistency.

New events are now explicitly set to Busy.  Existing events that
already have a BusyStatus are left alone.  (There is no UI for the
user to pick a different BusyStatus, so new events will always be
Busy.)

Fix #1417