Get BodyCalendarView information from MeetingRequest instead of iCal
Change BodyCalendarView to get its information about the meeting from
the MeetingRequest element in the ActiveSync XML for the e-mail
message.  (The MeetingRequest element is slightly misnamed.  It is
also used for responses and cancelations.)  BodyCalendarView no longer
uses the iCal information from the body of the e-mail message.

This update is not complete.  The code is not fully functional, and
some code cleanup is still needed.