McEmailAddress.Get doesn't acquire write lock if item already exists
Change McEmailAddress.Get to look for an existing item without
acquiring the write lock.  If an existing item is not found, then it
acquires the lock, repeats the lookup, and creates a new
McEmailAddress item if necessary.  The purpose of this change is to
avoid blocking for long periods of time while on the UI thread.

Fix #1522