Don't crash when starting from a notification and migration is required
If the user started the app by swiping a notification and a database
migration is required as part of startup, then the app would crash
during startup.  The problem was that NcApplication.Instance.Account
was not initialized in time.  It wasn't initialized until after
migration completed, but it was referenced from the code that deals
with the notification, which is in AppDelegate.FinishedLaunching.

The fix is to move the initialization of
NcApplication.Instance.Account from StartBasalServicesCompletion
(which might be delayed until after migration) to StartBasalServices.

Fix #1543