Overhaul the management of McEvents.
Overhaul the code that creates and manages McEvents.  McEvents are now
created on a background thread that doesn't block either the UI thread
or the sync response processing.  Code that needs to wait for McEvents
to be updated should listen for the EventSetChanged status.

The code that gathers and sorts the McEvents for the calendar view now
runs on a background thread, so it doesn't block the UI thread and
freeze the calendar view.  The calendar view doesn't get populated any
faster, but it is now responsive to user interaction while the
computations are happening.

Fix #1560
Fix #1536
Fix #1531