Keep calendar view from scrolling when data is reloaded
When UITableView.ReloadData() causes new rows to be added or existing
rows to disappear, it is possible that the visible area of the view
will shift, so the user is suddenly looking at a different set of
days.  This was a significant problem, because it was happening
consistently when the calendar view was first opened.

Change CalendarViewController to scroll the view back to where it was
if ReloadData() caused it shift significantly.