Improve the performance of McCalendar.Delete()
When a McCalendar item is deleted, all of its related exceptions and
events also need to be deleted.  This was being done by looping
through all of the exceptions and events and calling Delete() for each
one.  This proved to be inefficient.

The code was changed to use just one SQL DELETE command each for
exceptions and events, deleting them all at once instead of one at a
time.

Canceling the local notifications for the events being deleted
requires running some code on the UI thread.  Even if the UI thread
action is invoked asynchronously, the UI thread will take priority and
run for a little while, slowing down the background thread.  Move the
local notification cancelation code outside of the database
transaction.  This change doesn't make it run any faster, but it
avoids running it when an important lock is being held.