Remove some of the error messages about duplicate UIDs for events
Some servers will do an add and delete of a calendar item.  This
results in two items with the same UID in the database for a short
period of time.  Since duplicate UIDs are sometimes normal, and since
we can't detect the difference between a normal case and a buggy
situation, some of the error messages about duplicate UIDs have been
removed.

The message "Duplicate McCalendar Entry" is gone.  The message "Two
events have the same UID" was changed from an error to an
informational message.  The message "There are {0} events with the
same UID" remains an error, since that message happens when the
duplicate UIDs become a real problem rather than just a theoretical
problem.

Fix #945