Fix swiping a notification after a fresh install
If the app went through the login process or a database migration when
launched, then swiping an e-mail notification would not work.  It
would open the app, but not take the user to the message that was
swiped.  This is fixed by doing a more thorough search for the
NachoTabBarController.

Fix #1589