Don't crash when LocalNotificationManager sees an orphaned McEvent
It is possible for LocalNotificationManager to process a McEvent has
has been deleted or whose parent McException has been deleted.  The
local notification for that event will be canceled in very near future
as the normal McEvent processing proceeds.  But in the meantime,
LocalNotificationManager shouldn't crash when
McEvent.GetCalendarItemforEvent() returns null.

Fix nachocove/qa#74