Handle an ActiveSync time zone with no ID and no daylight saving time
Sometimes the time zone information in a calendar item doesn't have a
name, which causes problems when creating a C# time zone structure.
This situation was being handled correctly when the time zone had
daylight saving time.  But I had missed the case of a time zone with
no daylight saving time.  This change fixes that oversight.

Fix nachocove/qa#81