Fix nachocove/qa#95.
Break out of indexing loop if abatement is required. Change the scheduling of brain actions so that indexing gets to run. Specifically, split the remaining # of entries between indexing and score updates. Run indexing first so that it will not be constantly interrupted by hot view abatement signals triggered by score updates.