Open the correct message when tapping on a card in Hot view
There appears to be a bug in Xamarin or iOS where the indexPath that
is passed into HotListTableViewSource.RowSelected() is sometimes the
wrong row.  I haven't figured out the pattern of when the bug happens,
so the code has been changed to ignore the row that is passed in and
instead figure out which message is currently visible in the Hot view.

Fix #1336