Recognize exceptions that indicate a cancelled event occurrence
The Deleted element of an event exception is used to indicate that an
occurrence of a repeating event has been cancelled.  The code was
ignoring that field.  Fix the code to pay attention to the Deleted
field and to not create a McEvent for an event occurrence that has
been cancelled.

Fix nachocove/qa#71