Correctly display exceptions for repeating events
Fields in a McException are often left unset when they are the same as
the underlying meeting.  The app code for displaying events was not
taking this into account and was displaying the empty fields in the
calendar view and the event detail view.

Add methods to McAbstrCalendarRoot to access these fields.
McException overrides these methods to get the value either from
itself or from the main McCalendar.  Any code that accesses these
fields through a McAbstrCalendarRoot was changed to use the new
methods.  Code that accesses those fields through a McCalendar was
left alone.

Exceptions to recurring events should now display correctly in the
calendar view and in the event detail view.  (Mostly.  Attendees and
descriptions are sometimes wrong.  Those will be fixed separately.)

Fix nachocove/qa#69
Fix nachocove/qa#70